DeprecatedCloudFunctions/ImageFiltering/requirements.txt
google-cloud-storage
google-cloud-vision
Wand
firebase
firebase-admin
python_jwt
gcloud
sseclient
pycrypto
requests-toolbelt
google-cloud-language
datalab